Split task registration by plugin used
